Restructure to use common folder for components
