+++ packages/react-scripts/config/paths.js
@@ -50,10 +50,10 @@ module.exports = {
   appBuild: resolveApp('build/static'),
   appPublic: resolveApp('public'),
   appHtml: resolveApp('public/index.html'),
-  appIndexJs: resolveApp('app/client/index.js'),
+  appIndexJs: resolveApp('src/client/index.js'),
   appPackageJson: resolveApp('package.json'),
-  appSrc: resolveApp('app/client'),
-  serverIndexJs: resolveApp('app/server/index.js'),
+  appSrc: resolveApp('src'),
+  serverIndexJs: resolveApp('src/server/index.js'),
   yarnLockFile: resolveApp('yarn.lock'),
   testsSetup: resolveApp('src/setupTests.js'),
   appNodeModules: resolveApp('node_modules'),
@@ -69,10 +69,10 @@ if (__dirname.indexOf(path.join('packages', 'react-scripts', 'config')) !== -1) 
     appBuild: resolveOwn('../build/static'),
     appPublic: resolveOwn('../template/public'),
     appHtml: resolveOwn('../template/public/index.html'),
-    appIndexJs: resolveOwn('../template/app/client/index.js'),
+    appIndexJs: resolveOwn('../template/src/client/index.js'),
     appPackageJson: resolveOwn('../package.json'),
-    appSrc: resolveOwn('../template/app/client'),
-    serverIndexJs: resolveOwn('../template/app/server/index.js'),
+    appSrc: resolveOwn('../template/src'),
+    serverIndexJs: resolveOwn('../template/src/server/index.js'),
     yarnLockFile: resolveOwn('../template/yarn.lock'),
     testsSetup: resolveOwn('../template/src/setupTests.js'),
     appNodeModules: resolveOwn('../node_modules'),
