+++ packages/react-scripts/config/webpack.config.dev.js
@@ -81,6 +81,7 @@ module.exports = {
     // some tools, although we do not recommend using it, see:
     // https://github.com/facebookincubator/create-react-app/issues/290
     extensions: ['.js', '.json', '.jsx', '.scss'],
+    modules: [paths.appSrc, paths.ownNodeModules],
     alias: {
       // Support React Native Web
       // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
