+++ packages/react-scripts/template/src/client/index.js
@@ -1,7 +1,6 @@
-import './index.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from '../common/components/App';
 
 ReactDOM.render(
   <App />,
