+++ packages/react-scripts/template/src/common/components/App/App.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import logo from './logo.svg';
 import './App.css';
 
 const App = React.createClass({
@@ -7,15 +6,10 @@ const App = React.createClass({
     return (
       <div className="App">
         <div className="App-header">
-          <img
-            src={logo}
-            className="App-logo"
-            alt="logo"
-          />
           <h2>Welcome to React</h2>
         </div>
         <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
+          To get started, edit <code>src/common/App.js</code> and save to reload.
         </p>
       </div>
     );
