+++ packages/react-scripts/template/src/common/components/App/App.test.js
