+++ packages/react-scripts/template/src/server/index.js
