+++ packages/react-scripts/template/src/server/models/currencies.js
