+++ packages/react-scripts/template/src/server/routes/index.js
