+++ packages/react-scripts/template/src/server/routes/root.js
