+++ packages/react-scripts/template/src/server/system/index.js
