+++ packages/react-scripts/template/src/server/system/nunjucks.js
