+++ packages/react-scripts/template/src/server/views/application.html
