+++ packages/react-scripts/template/src/server/views/default.html
