+++ packages/react-scripts/template/src/server/views/error.html
